# lib/api-zod/src/generated/types/collection.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Facade order bot API
 * OpenAPI spec version: 0.1.0
 */

export interface Collection {
  id: number;
  name: string;
  manufacturerId: number;
  createdAt: Date;
  updatedAt: Date;
}