# lib/api-zod/src/generated/types/createOrderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Facade order bot API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItemInput } from "./orderItemInput";

export interface CreateOrderBody {
  regionId: number;
  decorId: number;
  customerName: string;
  customerContact: string;
  items: OrderItemInput[];
  attachedFileUrl?: string;
}