# lib/api-zod/src/generated/types/updatePriceBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Facade order bot API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdatePriceBody {
  pricePerSqm?: string;
  pricePerHole?: string;
  pricePackagingPerSqm?: string;
}